refactor(routes): update MongoDB import paths to use db.provider
