+++ server/routes/favorite-schedules/[id].delete.ts
@@ -1,4 +1,4 @@
-import { useMongoDB } from '../../provider/mongodb.provider'
+import { useMongoDB } from '../../provider/db.provider'
 
 export default defineEventHandler(async (event) => {
   const id =  getRouterParams(event).id
