+++ server/routes/schedules/[id].delete.ts
@@ -1,5 +1,5 @@
 import { ObjectId } from 'mongodb'
-import { useMongoDB } from '../../provider/mongodb.provider'
+import { useMongoDB } from '../../provider/db.provider'
 
 export default defineEventHandler(async (event) => {
   const { id } = getRouterParams(event)
